requirements.txt
ansicolors==1.0.2
beautifulsoup4==4.6.0
certifi==2017.4.17
chardet==3.0.3
configparser==3.5.0
Django==1.11.2
django-gulp==2.5.1
django-modelcluster==3.1
django-taggit==0.22.1
django-treebeard==4.1.1
django-webpack-loader==0.5.0
djangorestframework==3.6.3
env-tools==2.0.0
futures==3.0.5
html5lib==0.999999999
idna==2.5
olefile==0.44
optional-django==0.3.0
Pillow==4.1.1
psutil==4.3.0
psycopg2==2.7.1
pytz==2017.2
requests==2.17.3
six==1.10.0
tini==3.0.1
Unidecode==0.4.20
urllib3==1.21.1
wagtail==1.10.1
webencodings==0.5.1
Willow==0.4
